+++ app/src/main/res/values/colors.xml
@@ -9,5 +9,4 @@
     <color name="accepted">#EB00FF</color>
     <color name="borrowed">#FD0505</color>
     <color name="tempPhotoBackground">#C4C4C4</color>
-    <color name="black">#000000</color>
 </resources>
